+++ packages/vue3/src/components/DateInput/DateInput.vue
@@ -1,147 +1,75 @@
 <script setup lang="ts">
-import type { DayStatus } from '@youcan/ui-core';
-import { DateUtils } from '@youcan/ui-core';
+import { onClickOutside } from '@vueuse/core';
 import { computed, ref } from 'vue';
-import Day from './Internal/Day.vue';
-import MonthSwitcher from './Internal/MonthSwitcher.vue';
+import { DateUtils } from '@youcan/ui-core';
+import DatePicker from './DatePicker.vue';
 import type { DateInputValue } from './types';
+import SecondaryButton from '~/components/Button/SecondaryButton.vue';
 
 const props = defineProps<{
   modelValue: DateInputValue
+  disabled: boolean
 }>();
 
 const emit = defineEmits(['update:modelValue']);
 
 const month = ref(new Date());
-const today = computed(() => month.value);
 const model = computed({
   get: () => props.modelValue,
   set: (value: DateInputValue | null) => emit('update:modelValue', value),
 });
 
-const days = computed(() => DateUtils.getDaysForDateMonthCycle(today.value));
-const weekDays = DateUtils.weekDays();
-const lockMouseHover = ref(false);
-
-const handleDayClick = (day: DayStatus) => {
-  lockMouseHover.value = !lockMouseHover.value;
-
-  if (!model.value?.start || model.value?.end) {
-    if (!lockMouseHover.value) {
-      return;
-    }
-
-    model.value = {
-      start: day.date,
-      end: undefined,
-    };
-
-    return;
-  }
-
-  if (!model.value?.end) {
-    const endIsBiggerThanStartDate = DateUtils.biggerThan(model.value.start, day.date);
-
-    model.value = {
-      start: endIsBiggerThanStartDate ? day.date : model.value.start,
-      end: endIsBiggerThanStartDate ? model.value.start : day.date,
-    };
+const isDatePickerVisible = ref(false);
+const datePicker = ref<HTMLDivElement>();
+const toggleDatePicker = (state = !isDatePickerVisible.value) => {
+  if (!props.disabled) {
+    isDatePickerVisible.value = state;
   }
 };
 
-const handleDayHover = (day: DayStatus) => {
-  if (!model.value?.start || !lockMouseHover.value) {
-    return;
-  }
-
-  if (!model.value?.end || lockMouseHover.value) {
-    if (DateUtils.isSameDay(day.date, model.value.start)) {
-      return;
-    }
-
-    const endIsBiggerThanStartDate = DateUtils.biggerThan(model.value.start, day.date);
-
-    model.value = {
-      start: endIsBiggerThanStartDate ? day.date : model.value.start,
-      end: model.value.end && endIsBiggerThanStartDate ? model.value.end : (endIsBiggerThanStartDate ? model.value.start : day.date),
-    };
-  }
-};
-
-const getEdge = (day: DayStatus) => {
-  if (!model.value?.start && !model.value?.end) {
-    return day.isToday ? 'both' : 'none';
-  }
-
-  if (model.value.start && model.value.end) {
-    if (DateUtils.isSameDay(model.value.start, model.value.end)) {
-      return 'both';
-    }
-    else if (DateUtils.isBetween(day.date, model.value.start, model.value.end)) {
-      return 'middle';
-    }
-  }
-
-  if (model.value.start && DateUtils.isSameDay(day.date, model.value.start)) {
-    return 'start';
-  }
-
-  if (model.value.end && DateUtils.isSameDay(day.date, model.value.end)) {
-    return 'end';
-  }
-
-  if (model.value?.start && model.value?.end && day.isToday) {
-    return 'both';
-  }
-
-  return 'none';
-};
-
-const isActive = (day: DayStatus) => {
-  if (
-    (model.value?.start && DateUtils.isSameDay(day.date, model.value.start))
-    || (model.value?.end && DateUtils.isSameDay(day.date, model.value.end))) {
-    return true;
-  }
-
-  if (!model.value?.start || !model.value?.end) {
-    return false;
-  }
-
-  return DateUtils.isBetween(day.date, model.value.start, model.value.end);
-};
+onClickOutside(datePicker, () => toggleDatePicker(false));
 </script>
 
 <template>
   <div class="date-input">
-    <MonthSwitcher v-model="month" />
-    <div class="days-container">
-      <Day v-for="weekDay in weekDays" :key="weekDay.getTime()" :date="weekDay" edge="none" :is-selected="false"
-        :is-today="false" :disabled="true" :in-current-month="false" format="2-letters" />
-      <Day v-for="day in days" :key="day.date.getTime()" :date="day.date" :edge="getEdge(day)"
-        :is-selected="isActive(day)" :is-today="day.isToday" :disabled="false" :in-current-month="day.inCurrentMonth"
-        @click="handleDayClick(day)" @mouseover="handleDayHover(day)" />
+    <SecondaryButton size="sm" icon-position="right" class="input-trigger" :disabled="disabled"
+      @click="toggleDatePicker()">
+      <span v-show="model.start || model.start">
+        {{ DateUtils.getCalendarDay(model.start, 'Start') }} - {{ DateUtils.getCalendarDay(model.end, 'End') }}
+      </span>
+      <span v-show="!model.start && !model.start">Select date range</span>
+      <template #icon>
+        <i class="i-youcan-calendar" />
+      </template>
+    </SecondaryButton>
+    <div class="date-picker-container">
+      <DatePicker v-show="isDatePickerVisible" ref="datePicker" v-model="model" />
     </div>
   </div>
 </template>
 
 <style scoped>
 .date-input {
-  display: flex;
-  flex-direction: column;
-  width: max-content;
-  gap: 24px;
-  padding: 12px;
-  background-color: var(--base-white);
-  border-radius: 8px;
-  box-shadow: 0px 6px 6px 0px rgba(0, 0, 0, 0.06);
-  border: 1px solid #EBEBEB;
+  --width: 460px;
+
+  width: var(--width);
 }
 
-.days-container {
-  width: max-content;
-  display: grid;
-  grid-template-columns: repeat(7, auto);
-  background-color: var(--base-white);
+.input-trigger {
+  width: 100%;
+}
+
+.input-trigger {
+  justify-content: space-between;
+  --icon-color: gray(--gray-500);
+}
+
+.date-picker-container {
+  position: relative;
+}
+
+.date-picker-container .date-picker {
+  position: absolute;
+  top: 8px;
 }
 </style>
